+++ src/components/ProductSwitcher/ProductSwitcher.tsx
@@ -6,6 +6,7 @@ import './ProductSwitcher.less';
 interface IItem {
     title: string;
     value: string;
+    unit: string;
 }
 
 interface IProductSwitcherProps {
@@ -14,7 +15,7 @@ interface IProductSwitcherProps {
     /** Start index */
     startIndex?: number;
     /** Theme */
-    theme?: 'tariff-showcase' | 'with-unit-values';
+    theme?: 'tariff-showcase' | 'with-unit-values' | 'with-color-row';
     /** Custom class name */
     className: string;
     /** Change handler */
@@ -24,6 +25,7 @@ interface IProductSwitcherProps {
 interface IProductSwitcherState {
     /** Current value */
     currentValue: string;
+    currentIndex: number;
 }
 
 const cn = cnCreate('mfui-product-switcher');
@@ -32,7 +34,7 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         items: PropTypes.array.isRequired,
         startIndex: PropTypes.number,
         onChange: PropTypes.func,
-        theme: PropTypes.oneOf(['tariff-showcase', 'with-unit-values']),
+        theme: PropTypes.oneOf(['tariff-showcase', 'with-unit-values', 'with-color-row']),
     };
 
     static defaultProps: Pick<IProductSwitcherProps, 'startIndex'> = {
@@ -42,6 +44,7 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
     labelNodes: object;
     rootNode: HTMLElement | null;
     pointerNode: HTMLElement | null;
+    colorRowNode: HTMLElement | null;
     timer: number;
 
     constructor(props: IProductSwitcherProps) {
@@ -52,7 +55,7 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
 
         this.labelNodes = {};
 
-        this.state = { currentValue: items[safeStartIndex].value };
+        this.state = { currentValue: items[safeStartIndex].value, currentIndex: safeStartIndex };
     }
 
     componentDidMount() {
@@ -81,7 +84,7 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
             return;
         }
 
-        this.setState({ currentValue: value });
+        this.setState({ currentValue: value, currentIndex: index });
         this.movePointer(value);
     }
 
@@ -93,6 +96,10 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
 
         if (this.pointerNode) {
             this.pointerNode.style.transform = `translateX(${offsetValue}px)`;
+        }
+
+        if (this.colorRowNode) {
+            this.colorRowNode.style.width = `${offsetValue}px`;
         }
     }
 
@@ -108,6 +115,10 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         this.pointerNode = node;
     }
 
+    getColorRowNode = (node: HTMLElement | null) => {
+        this.colorRowNode = node;
+    }
+
     render() {
         const { items, theme, className } = this.props;
 
@@ -116,17 +127,24 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
                 <div className={cn('row')}>
                     {items.map((item, index) =>
                         <div
-                            className={cn('item', { active: item.value === this.state.currentValue })}
+                            className={cn('item',
+                                            {
+                                                active: item.value === this.state.currentValue,
+                                                disabled: index < this.state.currentIndex,
+                                            }
+                                        )}
                             key={item.title + item.value}
                             onClick={this.handleClickItem(item.value, index)}
                         >
                             <div className={cn('label')} ref={this.getLabelNodes(item)}>
                                 {item.title}
+                                {item.unit && <span className={cn('label-unit')}>{item.unit}</span>}
                             </div>
                         </div>
                     )}
                 </div>
-                <div className={cn('pointer')} ref={this.getPointerNode} />
+                <div className={cn('pointer')} ref={this.getPointerNode}/>
+                <div className={cn('color-row')} ref={this.getColorRowNode}/>
             </div>
         );
     }